+++ packages/grafana-toolkit/package.json
@@ -12,17 +12,18 @@
   ],
   "repository": {
     "type": "git",
-    "url": "http://github.com/grafana/grafana.git"
+    "url": "http://github.com/grafana/grafana.git",
+    "directory": "packages/grafana-toolkit"
   },
   "bin": {
     "grafana-toolkit": "./bin/grafana-toolkit.js"
   },
   "scripts": {
-    "lint": "eslint src/ --ext=.js,.ts,.tsx",
-    "typecheck": "tsc --noEmit",
-    "precommit": "npm run lint & npm run typecheck",
+    "build": "grafana-toolkit toolkit:build",
     "clean": "rimraf ./dist ./compiled",
-    "build": "grafana-toolkit toolkit:build"
+    "lint": "eslint src/ --ext=.js,.ts,.tsx",
+    "precommit": "npm run lint & npm run typecheck",
+    "typecheck": "tsc --noEmit"
   },
   "main": "src/index.ts",
   "dependencies": {
@@ -31,6 +32,7 @@
     "@grafana/data": "6.7.0-pre",
     "@grafana/eslint-config": "^1.0.0-rc1",
     "@grafana/tsconfig": "^1.0.0-rc1",
+    "@grafana/ui": "6.7.0-pre",
     "@types/command-exists": "^1.2.0",
     "@types/execa": "^0.9.0",
     "@types/expect-puppeteer": "3.3.1",
