+++ packages/grafana-toolkit/src/cli/utils/githubRelease.ts
@@ -1,5 +1,8 @@
-import { execLine, getPluginVersion } from './execLine';
+import { execLine } from './execLine';
 import { getPluginId } from '../../config/utils/getPluginId';
+import { getPluginJson } from '../../config/utils/pluginValidation';
+import { getCiFolder } from '../../plugins/env';
+import path = require('path');
 
 const ghrPlatform = (): string => {
   switch (process.platform) {
@@ -18,18 +21,19 @@ const GHR_VERSION = '0.13.0';
 const GHR_ARCH = process.arch === 'x64' ? 'amd64' : '386';
 const GHR_PLATFORM = ghrPlatform();
 const GHR_EXTENSION = process.platform === 'linux' ? 'tar.gz' : 'zip';
+const PUBLISH_DIR = path.resolve(getCiFolder(), 'packages');
 
 class GitHubRelease {
   token: string;
   username: string;
-  circleSha1: string;
+  commitHash: string;
   releaseNotes: string;
-  constructor(token: string, username: string, circleSha1: string, releaseNotes: string) {
+  constructor(token: string, username: string, commitHash: string, releaseNotes: string) {
     // Get the ghr binary according to platform
     this.getGhr();
     this.token = token;
     this.username = username;
-    this.circleSha1 = circleSha1;
+    this.commitHash = commitHash;
     this.releaseNotes = releaseNotes;
   }
 
@@ -49,16 +53,21 @@ class GitHubRelease {
   }
 
   release() {
+    const ciDir = getCiFolder();
+    const distDir = path.resolve(ciDir, 'dist');
+    const distContentDir = path.resolve(distDir, getPluginId());
+    const pluginJsonFile = path.resolve(distContentDir, 'plugin.json');
+    const pluginJson = getPluginJson(pluginJsonFile);
     execLine(`ghr
       -t "${this.token}"
       -u "${this.username}"
       -r "${getPluginId()}"
-      -c "${this.circleSha1}"
-      -n "${getPluginId()}_v${getPluginVersion()}"
+      -c "${this.commitHash}"
+      -n "${getPluginId()}_v${pluginJson.info.version}"
       -b "${this.releaseNotes}"
       -delete
-      "v${getPluginVersion()}"
-      ./artifacts/`);
+      "v${pluginJson.info.version}"
+      ${PUBLISH_DIR}`);
   }
 }
